Rock,Paper and Scissors
This is a rock-paper-scissors game where it is the player vs the computer. The computer’s answer will be randomly generated, while the program will ask the user for their input.The first one to get 5 points wins.
This project will better your understanding of while loops and if statements.